dashboard/requirements.txt
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
plotly==5.21.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
